+++ tests/pat-info.test.js
@@ -7,7 +7,7 @@ dotenv.config();
 import { jest } from "@jest/globals";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
-import patInfo, { RATE_LIMIT_SECONDS } from "../api/status/pat-info.js";
+import patInfo, { RATE_LIMIT_SECONDS } from "../api-renamed/status/pat-info.js";
 import { expect, it, describe, afterEach, beforeAll } from "@jest/globals";
 
 const mock = new MockAdapter(axios);
