+++ tests/status.up.test.js
@@ -4,7 +4,7 @@
 import { jest } from "@jest/globals";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
-import up, { RATE_LIMIT_SECONDS } from "../api/status/up.js";
+import up, { RATE_LIMIT_SECONDS } from "../api-renamed/status/up.js";
 import { expect, it, describe, afterEach } from "@jest/globals";
 
 const mock = new MockAdapter(axios);
