+++ tests/top-langs.test.js
@@ -2,7 +2,7 @@ import { jest } from "@jest/globals";
 import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
-import topLangs from "../api/top-langs.js";
+import topLangs from "../api-renamed/top-langs.js";
 import { renderTopLanguages } from "../src/cards/top-languages-card.js";
 import { renderError } from "../src/common/utils.js";
 import { expect, it, describe, afterEach } from "@jest/globals";
